requirements.txt
selenium
pytest
pytest-xdist
requests
pyyaml